updated hybridization approach
