+++ src/gmri2fem/mixed_t1map.py
@@ -9,7 +9,7 @@ import scipy
 import skimage
 from numpy.lib.stride_tricks import sliding_window_view
 
-from gmri2fem.utils import create_csf_mask
+from gmri2fem.masking import create_csf_mask
 from simple_mri import load_mri
